requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
channels
channels-redis
langchain
langchain-google-genai
chromadb==0.5.0  
celery
redis
psycopg2-binary
pypdf
python-dotenv
gunicorn
django-cors-headers
langchain-community==0.0.20
daphne==4.0  